restructured logic code, added support for custom logic actions, added support for better map gen, added trees and cut tree action

+++ Pather.Common/Definitions/AStar/AStarGraph.cs
@@ -7,7 +7,7 @@ namespace Pather.Common.Definitions.AStar
     [IgnoreNamespace]
     public class AStarGraph
     {
-        public AStarGraph(int[][] gridWeight)
+        public AStarGraph(double[][] gridWeight)
         {
         }
